backend/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
httpx
feedparser
readability-lxml
beautifulsoup4
lxml
orjson